refactor(ui): minor tab/layout tweaks (#1760)
