+++ ui/src/features/freightline/freight-item.tsx
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 import { Freight } from '@ui/gen/v1alpha1/generated_pb';
 
 import { FreightLabel } from '../common/freight-label';
-import { FreightMode } from '../project/project-details/types';
+import { FreightMode } from '../project/pipelines/types';
 
 import styles from './freightline.module.less';
