+++ ui/src/features/project/analysis-templates/analysis-templates-list.tsx
@@ -1,7 +1,7 @@
 import { useMutation, useQuery } from '@connectrpc/connect-query';
-import { faChartBar, faPlus, faTrash } from '@fortawesome/free-solid-svg-icons';
+import { faPlus, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { Button, Space, Table } from 'antd';
+import { Button, Table } from 'antd';
 import { format } from 'date-fns';
 import { useParams } from 'react-router-dom';
 
@@ -29,19 +29,6 @@ export const AnalysisTemplatesList = () => {
 
   return (
     <div className='p-4'>
-      <h1 className='pl-2 text-lg font-semibold flex items-center mb-4'>
-        <FontAwesomeIcon icon={faChartBar} className='mr-2' />
-        Analysis Templates
-        <Button
-          type='primary'
-          className='ml-auto'
-          icon={<FontAwesomeIcon icon={faPlus} />}
-          onClick={() => showCreate()}
-        >
-          New
-        </Button>
-      </h1>
-
       <Table<AnalysisTemplate>
         dataSource={data?.analysisTemplates}
         pagination={{ hideOnSinglePage: true }}
@@ -58,10 +45,23 @@ export const AnalysisTemplatesList = () => {
         <Table.Column<AnalysisTemplate> title='Name' dataIndex={['metadata', 'name']} />
         <Table.Column<AnalysisTemplate>
           width={100}
+          title={
+            <div className='w-full'>
+              <Button
+                type='primary'
+                className='ml-auto text-xs font-semibold'
+                icon={<FontAwesomeIcon icon={faPlus} />}
+                onClick={() => showCreate()}
+              >
+                ADD TEMPLATE
+              </Button>
+            </div>
+          }
           render={(_, template) => (
-            <Space>
+            <div className='w-full flex'>
               <Button
                 icon={<FontAwesomeIcon icon={faTrash} />}
+                className='ml-auto'
                 danger
                 loading={isDeleting}
                 onClick={() => {
@@ -86,7 +86,7 @@ export const AnalysisTemplatesList = () => {
               >
                 Delete
               </Button>
-            </Space>
+            </div>
           )}
         />
       </Table>
