+++ ui/src/features/project/pipelines/images.tsx
