+++ ui/src/features/project/pipelines/nodes/stage-node.module.less
