+++ ui/src/features/project/pipelines/project-details.module.less
