+++ ui/src/features/project/pipelines/utils/stage-yaml-example.ts
