+++ ui/src/features/project/pipelines/utils/warehouse-yaml-example.ts
